@@ -1,60 +1,22 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:flutter_translate_app/src/presentation/pages/home_page/home_page.dart';
-import 'package:flutter_translate_app/src/presentation/providers/database_provider.dart';
-import 'package:flutter_translate_app/src/presentation/providers/home_page_provider.dart';
-import 'package:flutter_translate_app/src/presentation/providers/language_provider.dart';
-import 'package:flutter_translate_app/src/presentation/providers/translate_provider.dart';
-import 'package:provider/provider.dart';
 
 void main() {
-  WidgetsFlutterBinding.ensureInitialized();
-
-  /// force portrait mode
-  SystemChrome.setPreferredOrientations(
-      [DeviceOrientation.portraitUp, DeviceOrientation.portraitDown]);
-
-  /// force transparent top bar
-  SystemChrome.setSystemUIOverlayStyle(const SystemUiOverlayStyle(
-    statusBarColor: Colors.transparent,
-  ));
-  runApp(MultiProvider(
-    providers: [
-      ChangeNotifierProvider(create: (_) => HomePageProvider()),
-      ChangeNotifierProvider(create: (_) => TranslateProvider()),
-      ChangeNotifierProvider(create: (_) => LanguageProvider()),
-      ChangeNotifierProvider(create: (_) => DatabaseProvider()),
-    ],
-    child: const MyApp(),
-  ));
+  runApp(const MyApp());
 }
 
-class MyApp extends StatefulWidget {
-  const MyApp({Key? key}) : super(key: key);
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
 
-  @override
-  State<MyApp> createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
+  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return ScreenUtilInit(
-      designSize: const Size(1080, 1920),
-      minTextAdapt: true,
-      splitScreenMode: true,
-      builder: (context, child) => MaterialApp(
-          title: 'Flutter translate app',
-          debugShowCheckedModeBanner: false,
-          theme: ThemeData(primarySwatch: Colors.red),
-          home: NotificationListener<OverscrollIndicatorNotification>(
-            onNotification: (overscroll) {
-              overscroll.disallowIndicator();
-              return false;
-            },
-            child: const HomePage(),
-          )),
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: Scaffold(),
     );
   }
 }
